バリデーション追加
