--- src/main/java/com/example/dai7kadai/RESTAPI/NotBlank.java
@@ -1,4 +1,0 @@
-package com.example.dai7kadai.RESTAPI;
-
-public @interface NotBlank {
-}
