+++ src/main/java/com/example/dai7kadai/restapi/CreateForm.java
@@ -1,0 +1,39 @@
+package com.example.dai7kadai.restapi;
+
+import jakarta.validation.constraints.Pattern;
+import org.hibernate.validator.constraints.Length;
+import org.springframework.validation.annotation.Validated;
+import jakarta.validation.constraints.NotBlank;
+
+@Validated
+public class CreateForm {
+    @NotBlank
+    @Length(max = 20)
+    private String name;
+
+    @NotBlank
+    @Pattern(regexp = "^19|20[0-9]{2}-(0[1-9]|1[0-2])-(0[1-9]|[12][0-9]|3[01])$")
+    private String birthday;
+
+    public CreateForm(String name, String birthday) {
+        this.name = name;
+        this.birthday = birthday;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public String getBirthday() {
+        return birthday;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public void setBirthday(String birthday) {
+        this.birthday = birthday;
+    }
+
+}
