+++ src/main/java/com/example/dai7kadai/restapi/Dai7kadaiRestapiApplication.java
@@ -1,4 +1,4 @@
-package com.example.dai7kadai.RESTAPI;
+package com.example.dai7kadai.restapi;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
